requirements.txt
pandas
matplotlib
seaborn
opencv-python
scikit-learn
ultralytics
Pillow
